gh-1426 refactoring
