+++ Multisig/UI/WalletConnect/Client/WCPendingConfirmationViewController.xib
