+++ Multisig/UI/WalletConnect/Server/Keys/WalletConnectKeysServerController.swift
@@ -1,0 +1,9 @@
+//
+//  WalletConnectKeysServerController.swift
+//  Multisig
+//
+//  Created by Andrey Scherbovich on 08.09.21.
+//  Copyright © 2021 Gnosis Ltd. All rights reserved.
+//
+
+import Foundation
